requirements.txt
pydicom
pillow
numpy
pymitter